Directory restructring: moved most files to src (including pairixmodule.c). Adjusted setup.py. Added MANIFEST.
