+++ src/bam_endian.h
