+++ src/khash.h
