+++ src/kseq.h
