+++ src/kstring.c
