+++ src/kstring.h
